+++ Runtime/Utils/Math/Extensions/mathx.cs
@@ -18,27 +18,23 @@ All rights reserved.
 */
 
 using System.Runtime.CompilerServices;
-using UnityEngine;
 using Unity.Mathematics;
 
 namespace Voxell.Mathx
 {
-  public static class float2_extension
+  public static class mathx
   {
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static float2 perpendicular(this float2 vector) => new float2(-vector.y, vector.x);
-
-    // Returns the angle in degrees between /from/ and /to/.
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static float angle(this float2 from, float2 to)
+    public static bool approximately(float a, float b)
     {
-      // sqrt(a) * sqrt(b) = sqrt(a * b) -- valid for real numbers
-      float denominator = (float)math.sqrt(math.lengthsq(from) * math.lengthsq(to));
-      if (denominator < Vector2.kEpsilonNormalSqrt)
-        return 0F;
-
-      float dot = math.clamp(math.dot(from, to) / denominator, -1F, 1F);
-      return math.acos(dot) * Mathf.Rad2Deg;
+      // If a or b is zero, compare that the other is less or equal to epsilon.
+      // If neither a or b are 0, then find an epsilon that is good for
+      // comparing numbers at the maximum magnitude of a and b.
+      // Floating points have about 7 significant digits, so
+      // 1.000001f can be represented while 1.0000001f is rounded to zero,
+      // thus we could use an epsilon of 0.000001f for comparing values close to 1.
+      // We multiply this epsilon by the biggest magnitude of a and b.
+      return math.abs(b - a) < math.max(0.000001f * math.max(math.abs(a), math.abs(b)), math.EPSILON * 8);
     }
   }
 }