+++ CHANGELOG.md
@@ -14,6 +14,10 @@
   - float2 extensions
     - float2 perpendicular vector
     - angle between 2 float2 vectors
+  - float3 extensions
+    - float3 LongAxis (select axis with the largest value)
+  - mathx
+    - approximately using Unity.Mathematics instead of Mathf
 
 ## [1.1.0]